refactor(compute): extract job workflow owner
