+++ scripts/ci/module-impact.json
@@ -124,6 +124,7 @@
     "compute_service": {
       "ownerPaths": [
         "src/main/compute/compute-host-profile-owner.ts",
+        "src/main/compute/compute-job-workflow-owner.ts",
         "src/main/compute/compute-remote-operation-owner.ts",
         "src/main/compute/compute-service.ts"
       ],
@@ -132,12 +133,15 @@
       "testFiles": {
         "owner": [
           "src/main/compute/compute-host-profile-owner.test.ts",
+          "src/main/compute/compute-job-workflow-owner.test.ts",
           "src/main/compute/compute-remote-operation-owner.test.ts",
           "src/main/compute/compute-service.test.ts",
           "src/main/compute/compute-jobs.integration.test.ts",
           "src/main/compute/concurrency-integration.test.ts"
         ],
         "contract": [
+          "src/main/compute/concurrency-manager.test.ts",
+          "src/main/compute/job-dispatcher.test.ts",
           "src/main/compute/scp-runner.test.ts",
           "src/main/compute/ssh-runner.test.ts",
           "src/main/compute/ipc.test.ts",
